avcodec/ccaption_dec: ignore leading non-breaking spaces
In Closed Captions (US), the non-breaking space (0xA0) can be used to
align text horizontally from the left when used as a leading character.
However, CC decoder does not ignore it as a leading character
like it does an ordinary space, so a blank padding is rendered
on the black CC box. This is not the intended viewing experience.

Ignore the leading non-breaking spaces, thus creating the intended
transparency which aligns the text. Since all characters are
fixed-width in CC, it can be handled the same way as we currently
treat leading ordinary spaces.

Also, as a nit, lowercase the NBSP's hex code in the entry table to match
casing of the other hex codes.

Signed-off-by: Marth64 <marth64@proxyid.net>